requirements.txt
gradio_client==0.16.1
python-telegram-bot==13.13
